requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
crypto-utils==1.0.0
idna==3.10
pycrypto==2.6.1
pycryptodome==3.21.0
requests==2.32.3
scapy==2.6.1
urllib3==2.3.0
